requirements.txt
# generated from manifests external_dependencies
email-validator
erpbrasil.assinatura
erpbrasil.base
erpbrasil.base>=2.3.0
erpbrasil.edoc
erpbrasil.edoc.pdf
erpbrasil.transmissao
mdfelib.v3_00
nfelib
nfselib.paulistana
num2words
phonenumbers
pycep_correios
pynfe.processamento.comunicacao
pynfe.utils
pyyaml
satcomum
unidecode
workalendar
